+++ src/main/java/chess/controller/state/Ready.java
@@ -1,4 +1,4 @@
-package chess.state;
+package chess.controller.state;
 
 import chess.domain.board.Board;
 import chess.domain.board.BoardFactory;
